product_suggest_controller追加しsuggest機能を移動、その他テストの修正
