+++ config/routes.rb
@@ -7,7 +7,7 @@ Rails.application.routes.draw do
   mount Spree::Core::Engine, at: '/'
 
   namespace :potepan do
-    get 'suggest', to: 'static_pages#suggest'
+    resources :product_suggests, only: [:index]
     resources :products, only: [:show]
     resources :categories, only: [:show]
     root 'static_pages#home'
